Gnatt

Documentation for Gnatt.

Gnatt.Job

Gnatt.gantiffy

Gnatt.yrmo

Gnatt.Job — Type

Job

Job wraps a start date, stop date, job name, job lead, and any relevant details. The Job type is mutable and accepts initial keyword definitions (e.g. Job(start=(2000,01))).

Gnatt.gantiffy — Method

gantiffy(jobs...)

Provides inputs to Makie's barplot function. Returns a NamedTuple that should be used as shown below:

Example

julia> g = gantiffy(job1, job2, job3, ...)

julia> f = Makie.Figure()

julia> ax = Makie.Axis(f[1,1]; g.axis...) 

julia> Makie.barplot!(ax,g.jobs, g.stops; g.kwargs...)

Note the use of the semicolon ; operators to separate args from kwargs in the Axis and barplot functions.

Add a legend

The example below shows how to add a legend above the plot.

julia> Legend(f[0,1], ax, tellheight=true, tellwidth=false, orientation=:horizontal);f

Gnatt.yrmo — Method

Gnatt.yrmo(d::Date)

Convert a year-month Date to float where the month is the ¹/₁₂ decimal of the year.

Examples

julia> Gnatt.yrmo(Date(1996,8))
1996.5833333333333